More package refactoring
Further organization of packages
