+++ src/test/java/asd/arch/GameLoopTest.java
@@ -1,4 +1,4 @@
-package asd;
+package asd.arch;
 
 import org.junit.jupiter.api.Test;
